build: use current commit if forcing publish
If there is a problem publishing the new version it may be required to
trigger the publish again without further incrementing the version.
Changing the `force-publish` input flag to instead accept a string
version tag allows a targetted version to be retried without any side
effects. New commits won't cause a version increment and won't be
included in the published build either.

NO-TICKET